% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{canDeleteRow}
\alias{canDeleteRow}
\title{Determine if a row can be edited}
\usage{
canDeleteRow(row, statusCol)
}
\arguments{
\item{row}{\code{data.frame}}

\item{statusCol}{\code{character(1)} name of column with general status (e.g. modified or not).}
}
\value{
\code{boolean}
}
\description{
Determine if a row can be edited
}
\author{
Jasper Schelfhout
}
