% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{canEditRow}
\alias{canEditRow}
\title{Determine if a row can be deleted}
\usage{
canEditRow(row, statusCol)
}
\arguments{
\item{row}{\code{data.frame}}

\item{statusCol}{\code{character(1)} name of column with general status (e.g. modified or not).}
}
\value{
\code{boolean}
}
\description{
Determine if a row can be deleted
}
\author{
Jasper Schelfhout
}
