% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{castForDisplay}
\alias{castForDisplay}
\title{Cast columns in data.frame to editable types in datatable}
\usage{
castForDisplay(data, cols = colnames(data))
}
\arguments{
\item{data}{data.frame}

\item{cols}{character columns perform casting on.}
}
\value{
data.frame with some columns cast to another type
}
\description{
Cast columns in data.frame to editable types in datatable
}
\author{
Jasper Schelfhout
}
