% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreignTbl.R
\name{castToFactor}
\alias{castToFactor}
\title{Cast all columns that exist in a foreignTbl to factor}
\usage{
castToFactor(data, foreignTbls)
}
\arguments{
\item{data}{\code{data.frame}}

\item{foreignTbls}{list of foreign tbls as created by \code{\link{foreignTbl}}}
}
\value{
data.frame
}
\description{
Cast all columns that exist in a foreignTbl to factor
}
\details{
Can be used to fixate possible options when editing.
}
\author{
Jasper Schelfhout
}
