% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dbi.R
\name{castToSQLSupportedType}
\alias{castToSQLSupportedType}
\title{Cast the data type to something supported by SQL.}
\usage{
castToSQLSupportedType(x)
}
\arguments{
\item{x}{single value or vector of values}
}
\value{
x, possibly cast to different type
}
\description{
Cast the data type to something supported by SQL.
}
\author{
Jasper Schelfhout
}
