% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{castToTemplate}
\alias{castToTemplate}
\title{Cast tbl or data.frame x to the types of the template}
\usage{
castToTemplate(x, template)
}
\arguments{
\item{x}{`data.frame`, `tbl` or `data.table`}

\item{template}{`data.frame`, `tbl` or `data.table`}
}
\value{
object containing data of x in the class and structure of the template.
}
\description{
Cast tbl or data.frame x to the types of the template
}
\details{
If template is a tbl with database source, convert to an in-memory tibble with same data types instead.

Rownames might show differing
}
\author{
Jasper Schelfhout
}
