% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreignTbl.R
\name{checkForeignTbls}
\alias{checkForeignTbls}
\title{Check if all rows in tbl fufill foreignTbl constraints}
\usage{
checkForeignTbls(tbl, foreignTbls)
}
\arguments{
\item{tbl}{tibble}

\item{foreignTbls}{list}
}
\value{
boolean if tbl fufills constraint of all foreign tbls.
}
\description{
Check if all rows in tbl fufill foreignTbl constraints
}
\author{
Jasper Schelfhout
}
