% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coalesce}
\alias{coalesce}
\title{Return first non null argument}
\usage{
coalesce(...)
}
\arguments{
\item{...}{set of arguments}
}
\description{
Return first non null argument
}
\author{
Jasper Schelfhout
}
