% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerceColumns}
\alias{coerceColumns}
\title{cast columns to the type of the template}
\usage{
coerceColumns(template, x)
}
\arguments{
\item{template}{data.frame}

\item{x}{data.frame}
}
\description{
cast columns to the type of the template
}
\details{
only affects columns in both the template and x
}
