% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerceValue}
\alias{coerceValue}
\title{DT coerceValue with better POSIXct support}
\usage{
coerceValue(val, old)
}
\arguments{
\item{val}{A character string.}

\item{old}{An old value, whose type is the target type of \code{val}.}
}
\description{
DT coerceValue with better POSIXct support
}
\author{
Jasper Schelfhout
}
