% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connectDB}
\alias{connectDB}
\title{Connect to your database.}
\usage{
connectDB(
  dbname = system.file("extdata", "chinook.db", package = utils::packageName()),
  drv = RSQLite::SQLite(),
  ...
)
}
\arguments{
\item{dbname}{\code{character(0)}}

\item{drv}{database driver}

\item{...}{arguments passed to \code{DBI::dbConnect}}
}
\value{
database connection
}
\description{
Connect to your database.
}
