% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{createButtons}
\alias{createButtons}
\title{Create buttons to modify the row.}
\usage{
createButtons(
  row,
  suffix,
  ns,
  canEditRow = editbl::canEditRow,
  canDeleteRow = editbl::canDeleteRow,
  statusCol = "status"
)
}
\arguments{
\item{row}{\code{tibble} with single row}

\item{suffix}{\code{character(1)}}

\item{ns}{\code{character(1)} namespace}

\item{canEditRow}{\code{function}.
Has function argument \code{row} which accepts a single row \code{tibble}.
Returns \code{logical(1)}.}

\item{canDeleteRow}{\code{function}.
Has function argument \code{row} which accepts a single row \code{tibble}.
Returns \code{logical(1)}.}

\item{statusCol}{\code{character(1)} name of column with general status (e.g. modified or not).}
}
\value{
\code{character(1)} HTML
}
\description{
Create buttons to modify the row.
}
\details{
buttons used per row in the app.
}
