% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{createDeleteButtonHTML_shiny}
\alias{createDeleteButtonHTML_shiny}
\title{Helper function to write HTML}
\usage{
createDeleteButtonHTML_shiny(ns = "\%1$s", suffix = "\%2$s", disabled = FALSE)
}
\arguments{
\item{ns}{\code{character(1)} namespace}

\item{suffix}{\code{character(1)} id of the row}

\item{disabled}{\code{logical(1)} wether or not the button has to be disabled}
}
\description{
Helper function to write HTML
}
\details{
only to be used interactively. sprintf() implementation
is faster.
}
\seealso{
createEditButtonHTML
}
