% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{createEditButtonHTML}
\alias{createEditButtonHTML}
\title{Helper function to write HTML}
\usage{
createEditButtonHTML(ns = "\%1$s", suffix = "\%2$s")
}
\arguments{
\item{ns}{\code{character(1)} sprintf placeholder for ns}

\item{suffix}{\code{character(1)} sprintf placeholer for suffix}
}
\value{
\code{character(1)} HTML to be filled in with \code{sprintf}
}
\description{
Helper function to write HTML
}
\details{
generate HTML as character once and reuse.
Since buttons have to be generated a lot, this otherwhise slows down the app.
}
