% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{customButton}
\alias{customButton}
\title{Generate a custom button for DT}
\usage{
customButton(id, label, icon = "", disabled = FALSE)
}
\arguments{
\item{id}{character, namespaced id}

\item{label}{character}

\item{icon}{shiny::icon}

\item{disabled}{boolean. Wether or not the button should start in a disabled state.}
}
\value{
list to be used in eDT(options = list(buttons = xxx))
}
\description{
Generate a custom button for DT
}
\details{
Combines elements of shiny and options defined here:
https://datatables.net/reference/option/
}
\author{
Jasper Schelfhout
}
