% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoApp.R
\name{demoServer_custom}
\alias{demoServer_custom}
\title{Server of the mtcars demo app}
\usage{
demoServer_custom(id, x)
}
\arguments{
\item{id}{\code{character(1)}}

\item{x}{\code{tbl}}
}
\description{
Server of the mtcars demo app
}
\author{
Jasper Schelfhout
}
