% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoApp.R
\name{demoServer_mtcars}
\alias{demoServer_mtcars}
\title{Server of the mtcars demo app}
\usage{
demoServer_mtcars(id)
}
\arguments{
\item{id}{\code{character(1)}}
}
\value{
NULL, just executes the module server.
}
\description{
Server of the mtcars demo app
}
\author{
Jasper Schelfhout
}
