% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoApp.R
\name{demoUI_custom}
\alias{demoUI_custom}
\title{UI of the demo mtcars app}
\usage{
demoUI_custom(id)
}
\arguments{
\item{id}{character}
}
\value{
HTML
}
\description{
UI of the demo mtcars app
}
\author{
Jasper Schelfhout
}
