% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoApp.R
\name{demoUI_DB}
\alias{demoUI_DB}
\title{UI of the DB demo app}
\usage{
demoUI_DB(id, conn)
}
\arguments{
\item{id}{\code{character(1)}}

\item{conn}{database connection object as given by \code{\link[DBI]{dbConnect}}.}
}
\value{
HTML
}
\description{
UI of the DB demo app
}
\author{
Jasper Schelfhout
}
