% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devApp.R
\name{devServer}
\alias{devServer}
\title{Server of the demo app}
\usage{
devServer(id, conn)
}
\arguments{
\item{id}{\code{character(1)}}

\item{conn}{database connection object as given by \code{\link[DBI]{dbConnect}}.}
}
\description{
Server of the demo app
}
\author{
Jasper Schelfhout
}
