% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devApp.R
\name{devUI}
\alias{devUI}
\title{UI of the development app}
\usage{
devUI(id, conn)
}
\arguments{
\item{id}{\code{character(1)}}

\item{conn}{database connection object as given by \code{\link[DBI]{dbConnect}}.}
}
\value{
HTML
}
\description{
UI of the development app
}
\author{
Jasper Schelfhout
}
