% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{disableDoubleClickButtonCss}
\alias{disableDoubleClickButtonCss}
\title{Function to generate CSS to disable clicking events on a column}
\usage{
disableDoubleClickButtonCss(id)
}
\arguments{
\item{id}{\code{character(1)} namespaced id of the datatable}
}
\value{
\code{character} CSS
}
\description{
Function to generate CSS to disable clicking events on a column
}
\details{
\url{https://stackoverflow.com/questions/60406027/how-to-disable-double-click-reactivity-for-specific-columns-in-r-datatable}

\url{https://stackoverflow.com/questions/75406546/apply-css-styling-to-a-single-dt-datatable}
}
