% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT_app.R
\name{eDT_app}
\alias{eDT_app}
\title{Open interactive app to explore and modify data}
\usage{
eDT_app(...)
}
\arguments{
\item{...}{arguments past to eDT}
}
\value{
data (or a modified version thereof)
}
\description{
Open interactive app to explore and modify data
}
