% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT_app.R
\name{eDT_app_server}
\alias{eDT_app_server}
\title{Server of eDT_app}
\usage{
eDT_app_server(moduleId = "nevergonnagiveyouup", ...)
}
\arguments{
\item{moduleId}{\code{character(1)} id to connect with eDT_app_server}

\item{...}{arguments passed to \link{eDT}}
}
\value{
moduleServer which on application stop returns version of x with made changes
}
\description{
Server of eDT_app
}
\author{
Jasper Schelfhout
}
