% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT_app.R
\name{eDT_app_ui}
\alias{eDT_app_ui}
\title{UI of eDT_app}
\usage{
eDT_app_ui(moduleId = "nevergonnagiveyouup", eDTId = "nevergonnaletyoudown")
}
\arguments{
\item{moduleId}{\code{character(1)} id to connect with eDT_app_server}

\item{eDTId}{\code{character(1)} id to connect \code{\link{eDTOutput}} to \code{\link{eDT}} within the module.}
}
\value{
HTML
}
\description{
UI of eDT_app
}
\author{
Jasper Schelfhout
}
