% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{eDTOutput}
\alias{eDTOutput}
\title{UI of the eDT module}
\usage{
eDTOutput(id, ...)
}
\arguments{
\item{id}{character}

\item{...}{arguments passed to  DT::DTOutput}
}
\value{
HTML
}
\description{
UI of the eDT module
}
\author{
Jasper Schelfhout
}
