% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{evalCanCloneRow}
\alias{evalCanCloneRow}
\title{Determine if a row can be cloned}
\usage{
evalCanCloneRow(row, canCloneRow = TRUE, statusCol = "_editbl_status")
}
\arguments{
\item{row}{\code{tibble}, single row.}

\item{canCloneRow}{\code{function} with argument 'row' defining logic on wether or
not the row can be cloned. Can also be \code{logical} TRUE or FALSE.}

\item{statusCol}{\code{character(1)} name of column with general status (e.g. modified or not).}
}
\value{
\code{boolean}
}
\description{
Determine if a row can be cloned
}
\details{
calling this around the user passed on function ensures
that newly inserted rows are being excempt from the logic.
Moreover, the output of the function can be checked.
}
\author{
Saar Junius
}
