% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{evalCanDeleteRow}
\alias{evalCanDeleteRow}
\title{Determine if a row can be deleted}
\usage{
evalCanDeleteRow(row, canDeleteRow = TRUE, statusCol = "status")
}
\arguments{
\item{row}{\code{tibble}, single row}

\item{canDeleteRow}{\code{function} with argument 'row' defining logic on wether or
not the row can be modified. Can also be \code{logical} TRUE or FALSE.}

\item{statusCol}{\code{character(1)} name of column with general status (e.g. modified or not).}
}
\value{
\code{boolean}
}
\description{
Determine if a row can be deleted
}
\details{
calling this around the user passed on function ensures
that newly inserted rows are being excempt from the logic.
Moreover, the output of the function can be checked.
}
\author{
Jasper Schelfhout
}
