% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dbi.R
\name{get_db_table_name}
\alias{get_db_table_name}
\title{get name of the tbl in the database}
\usage{
get_db_table_name(x)
}
\arguments{
\item{x}{tbl with DBI connection}
}
\value{
SQL
}
\description{
get name of the tbl in the database
}
