% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getColumnTypeSums}
\alias{getColumnTypeSums}
\title{Get types of columns in a tbl}
\usage{
getColumnTypeSums(tbl)
}
\arguments{
\item{tbl}{\code{tbl}}
}
\value{
named list with types of the colums
}
\description{
Get types of columns in a tbl
}
\author{
Jasper Schelfhout
}
