% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreignTbl.R
\name{getNonNaturalKeyCols}
\alias{getNonNaturalKeyCols}
\title{Get all columns that are not natural keys}
\usage{
getNonNaturalKeyCols(foreignTbls)
}
\arguments{
\item{foreignTbls}{list}
}
\value{
character
}
\description{
Get all columns that are not natural keys
}
\author{
Jasper Schelfhout
}
