% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInput.R
\name{inputServer}
\alias{inputServer}
\title{An input server for a \code{data.frame}}
\usage{
inputServer(id, data, ...)
}
\arguments{
\item{id}{\code{character(1)} module id}

\item{data}{single row \code{data.frame}}

\item{...}{further arguments for methods}
}
\value{
modified version of data
}
\description{
An input server for a \code{data.frame}
}
\examples{
\dontrun{
library(shiny)
ui <- inputUI('id')
server <- function(input,output,session){
  input <- inputServer("id", mtcars[1,])
  observe({print(input())})
}
shinyApp(ui, server)
}

}
\author{
Jasper Schelfhout
}
