% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInput.R
\name{inputUI.default}
\alias{inputUI.default}
\title{UI part for modal with input fields for editing}
\usage{
\method{inputUI}{default}(id, ...)
}
\arguments{
\item{id}{character module id}

\item{...}{for compatibility with method}
}
\value{
HTML. A set of input fields corresponding to the given row.
}
\description{
UI part for modal with input fields for editing
}
\details{
The UI elements that have an id identical to a column name are used for updating the data.
}
\author{
Jasper Schelfhout
}
