% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInput.R
\name{inputUI}
\alias{inputUI}
\title{An input UI for a data.frame}
\usage{
inputUI(id, ...)
}
\arguments{
\item{id}{character module id}

\item{...}{arguments passed onto methods}
}
\value{
HTML. A set of input fields corresponding to the given row.
}
\description{
An input UI for a data.frame
}
\author{
Jasper Schelfhout
}
