% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isMissing}
\alias{isMissing}
\title{Check if an argument is missing}
\usage{
isMissing(x)
}
\arguments{
\item{x}{object}
}
\value{
boolean
}
\description{
Check if an argument is missing
}
\details{
in contrast to `missing`, this can also be used in a nested environment
}
\author{
Jasper Schelfhout
}
