% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreignTbl.R
\name{joinForeignTbl}
\alias{joinForeignTbl}
\title{Merge a tbl with it a foreignTbl}
\usage{
joinForeignTbl(
  tbl,
  foreignTbl,
  keepNA = TRUE,
  by = foreignTbl$by,
  copy = TRUE,
  type = c("inner", "left")[1]
)
}
\arguments{
\item{tbl}{\code{tbl}}

\item{foreignTbl}{\code{list} as created by \code{\link{foreignTbl}}}

\item{keepNA}{\code{logical} keep rows from tbl with NA keys.}

\item{by}{named \code{character}}

\item{copy}{\code{logical}}

\item{type}{\code{character}}
}
\value{
tibble \code{tbl}
}
\description{
Merge a tbl with it a foreignTbl
}
\details{
see also \code{dplyr} join functions like for example \code{dplyr::left_join}
}
\examples{
\dontrun{
a <- tibble::as_tibble(data.frame(
    a = c(1,2,3),
    key1 = c(NA,2,4),
    key2 = c(NA,2,4)))

b <-  tibble::as_tibble(data.frame(
    b = c("a", "b", "c"),
    key1 = c(1,2,3),
    key2 = c(1,2,3)
))

foreignTbl <- foreignTbl(a,b)

joinForeignTbl(a, foreignTbl)

}

}
\author{
Jasper Schelfhout
}
