% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overwriteDefaults}
\alias{overwriteDefaults}
\title{Overwrite default settings with provided settings}
\usage{
overwriteDefaults(defaults, settings)
}
\arguments{
\item{defaults}{named character vector}

\item{settings}{named character vector}
}
\value{
named character vector
}
\description{
Overwrite default settings with provided settings
}
\author{
Jasper Schelfhout
}
