% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dbi.R
\name{rowDelete}
\alias{rowDelete}
\title{Delete rows from a table in the db}
\usage{
rowDelete(conn, table, where)
}
\arguments{
\item{conn}{database connection object as given by \code{\link[DBI]{dbConnect}}.}

\item{table}{character}

\item{where}{named list, values to filter on. Names are database column names. If NULL no filter is applied and all rows are deleted.}
}
\value{
integer number of affected rows.
}
\description{
Delete rows from a table in the db
}
