% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dbi.R
\name{rowInsert}
\alias{rowInsert}
\title{Add a row to a table in the database.}
\usage{
rowInsert(conn, table, values)
}
\arguments{
\item{conn}{database connection object as given by \code{\link[DBI]{dbConnect}}.}

\item{table}{character}

\item{values}{named list, row to add. Names are database column names. Unspecified columns will get database defaults.}
}
\value{
integer number of affected rows.
}
\description{
Add a row to a table in the database.
}
