% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dbi.R
\name{rowUpdate}
\alias{rowUpdate}
\title{Update rows in the database.}
\usage{
rowUpdate(conn, table, values, where)
}
\arguments{
\item{conn}{database connection object as given by \code{\link[DBI]{dbConnect}}.}

\item{table}{character}

\item{values}{named list, values to be set. Names are database column names.}

\item{where}{named list, values to filter on. Names are database column names. If NULL no filter is applied.}
}
\value{
integer number of affected rows.
}
\description{
Update rows in the database.
}
