% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoApp.R
\name{runDemoApp}
\alias{runDemoApp}
\title{Run a demo app}
\usage{
runDemoApp(app = "database", ...)
}
\arguments{
\item{app}{demoApp to run. Options: database / mtcars / custom}

\item{...}{arguments passed onto demoApp}
}
\description{
Run a demo app
}
\examples{
\dontrun{
 runDemoApp(app = "database")
 runDemoApp(app = "mtcars")
 runDemoApp(app = "custom", dplyr::tibble(iris))

}
}
