% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoApp.R
\name{runDemoApp_custom}
\alias{runDemoApp_custom}
\title{Run a custom demo app}
\usage{
runDemoApp_custom(x)
}
\arguments{
\item{x}{\code{tbl}}
}
\value{
An object that represents the app. Printing the object or passing it
to \code{\link[shiny:runApp]{runApp()}} will run the app.
}
\description{
Run a custom demo app
}
