% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoApp.R
\name{runDemoApp_DB}
\alias{runDemoApp_DB}
\title{Run a demo app}
\usage{
runDemoApp_DB()
}
\value{
An object that represents the app. Printing the object or passing it
to \code{\link[shiny:runApp]{runApp()}} will run the app.
}
\description{
Run a demo app
}
