% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInput.R
\name{selectInputDT_Server}
\alias{selectInputDT_Server}
\title{Server to use a datatable as select input}
\usage{
selectInputDT_Server(
  id,
  label = "",
  choices,
  selected = NULL,
  multiple = FALSE
)
}
\arguments{
\item{id}{character}

\item{label}{character}

\item{choices}{data.frame}

\item{selected}{data.frame}
}
\value{
data.frame
}
\description{
Server to use a datatable as select input
}
\author{
Jasper Schelfhout
}
