% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectInputDT.R
\name{selectInputDT_UI}
\alias{selectInputDT_UI}
\title{Create a DT select input}
\usage{
selectInputDT_UI(id)
}
\arguments{
\item{id}{\code{character(1)} same one as used in \code{\link{selectInputDT_Server}}}
}
\description{
Create a DT select input
}
\author{
Jasper Schelfhout
}
