% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInput.R
\name{shinyInput}
\alias{shinyInput}
\title{Get a shiny input for a column of a \code{tbl}}
\usage{
shinyInput(x, inputId, label, selected)
}
\arguments{
\item{x}{column}

\item{inputId}{shiny input Id}

\item{label}{\code{character(1)}}

\item{selected}{object of class of x}
}
\value{
shiny input
}
\description{
Get a shiny input for a column of a \code{tbl}
}
\author{
Jasper Schelfhout
}
