% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardizeColnamesArgument}
\alias{standardizeColnamesArgument}
\title{Standardize to colnames argument to the format of named character vector}
\usage{
standardizeColnamesArgument(x, data)
}
\arguments{
\item{x}{colnames argument to eDT}

\item{data}{data argument to eDT}
}
\value{
named character vector
}
\description{
Standardize to colnames argument to the format of named character vector
}
\author{
Jasper Schelfhout
}
