% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dbi.R
\name{whereSQL}
\alias{whereSQL}
\title{Generate where sql}
\usage{
whereSQL(conn, table, column, operator = "in", values = NULL)
}
\arguments{
\item{conn}{database connection object as given by \code{\link[DBI]{dbConnect}}.}

\item{table}{character table name (or alias used in query)}

\item{column}{character column of table}

\item{operator}{character}

\item{values}{character vector of values}
}
\value{
character sql
}
\description{
Generate where sql
}
\author{
Jasper Schelfhout
}
