% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{addButtons}
\alias{addButtons}
\title{Add modification buttons as a column}
\usage{
addButtons(df, columnName, ns)
}
\arguments{
\item{df}{\code{data.frame}}

\item{columnName}{\code{character(1)}}

\item{ns}{namespace function}
}
\value{
df with extra column containing buttons
}
\description{
Add modification buttons as a column
}
\author{
Jasper Schelfhout
}
