% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{beginTransaction}
\alias{beginTransaction}
\title{Start a transaction for a tibble}
\usage{
beginTransaction(tbl)
}
\arguments{
\item{tbl}{\code{tbl}}
}
\description{
Start a transaction for a tibble
}
\author{
Jasper Schelfhout
}
