% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{canXXXRowTemplate}
\alias{canXXXRowTemplate}
\title{Re-usable documentation}
\usage{
canXXXRowTemplate(canEditRow, canDeleteRow)
}
\arguments{
\item{canEditRow}{\code{function}.
Has function argument \code{row} which accepts a single row \code{tibble}.
Returns \code{logical(1)}.}

\item{canDeleteRow}{\code{function}.
Has function argument \code{row} which accepts a single row \code{tibble}.
Returns \code{logical(1)}.}
}
\description{
Re-usable documentation
}
