% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{canXXXRowTemplate}
\alias{canXXXRowTemplate}
\title{Re-usable documentation}
\usage{
canXXXRowTemplate(canEditRow, canDeleteRow)
}
\arguments{
\item{canEditRow}{can be either of the following:
\itemize{
\item \code{logical}, e.g. TRUE or FALSE
\item \code{function}. Needs as input an argument \code{row} which accepts a single row \code{tibble} and as output TRUE/FALSE.
}}

\item{canDeleteRow}{can be either of the following:
\itemize{
\item \code{logical}, e.g. TRUE or FALSE
\item \code{function}. Needs as input an argument \code{row} which accepts a single row \code{tibble} and as output TRUE/FALSE.
}}
}
\description{
Re-usable documentation
}
