% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{castForDisplay}
\alias{castForDisplay}
\title{Cast columns in \code{data.frame} to editable types in datatable}
\usage{
castForDisplay(data, cols = colnames(data))
}
\arguments{
\item{data}{\code{data.frame}}

\item{cols}{\code{character} columns to perform casting on.}
}
\value{
\code{data.frame} with some columns cast to another type
}
\description{
Cast columns in \code{data.frame} to editable types in datatable
}
\author{
Jasper Schelfhout
}
