% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{castFromTbl}
\alias{castFromTbl}
\title{Cast tbl to class of template}
\usage{
castFromTbl(tbl, template)
}
\arguments{
\item{tbl}{\code{tbl}}

\item{template}{tabular object like \code{data.frame} or \code{data.table} or \code{tbl}.}
}
\value{
tbl cast to the type of template
}
\description{
Cast tbl to class of template
}
\author{
Jasper Schelfhout
}
