% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreignTbl.R
\name{checkForeignTbls}
\alias{checkForeignTbls}
\title{Check if all rows in tbl fufill \code{foreignTbl} constraints}
\usage{
checkForeignTbls(tbl, foreignTbls)
}
\arguments{
\item{tbl}{\code{tbl}}

\item{foreignTbls}{list of foreign tbls as created by \code{\link{foreignTbl}}}
}
\value{
\code{logical} stating if tbl fufills all constraints imposed by all foreign tbls.
}
\description{
Check if all rows in tbl fufill \code{foreignTbl} constraints
}
\author{
Jasper Schelfhout
}
