% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerceColumns}
\alias{coerceColumns}
\title{Cast columns to the type of the template}
\usage{
coerceColumns(template, x)
}
\arguments{
\item{template}{\code{data.frame}}

\item{x}{\code{data.frame}}
}
\description{
Cast columns to the type of the template
}
\details{
only affects columns in both the template and x
}
