% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{createButtons}
\alias{createButtons}
\title{Create buttons to modify the row. See \code{\link{character}}}
\usage{
createButtons(suffix, ns)
}
\arguments{
\item{suffix}{character}

\item{ns}{namespace}
}
\value{
character HTML
}
\description{
Create buttons to modify the row. See \code{\link{character}}
}
\details{
buttons used per row in the app.
}
