% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{createButtons}
\alias{createButtons}
\title{Create buttons to modify the row. See \code{\link{createButtonsHTML}}}
\usage{
createButtons(suffix, ns)
}
\arguments{
\item{suffix}{\code{character(1)}}

\item{ns}{\code{character(1)} namespace}
}
\value{
\code{character} HTML
}
\description{
Create buttons to modify the row. See \code{\link{createButtonsHTML}}
}
\details{
buttons used per row in the app.
}
