% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{createButtons}
\alias{createButtons}
\title{Create buttons to modify the row.}
\usage{
createButtons(
  row,
  suffix,
  ns,
  canEditRow = TRUE,
  canDeleteRow = TRUE,
  canCloneRow = TRUE,
  statusCol = "_editbl_status"
)
}
\arguments{
\item{row}{\code{tibble} with single row}

\item{suffix}{\code{character(1)}}

\item{ns}{\code{character(1)} namespace}

\item{canEditRow}{can be either of the following:
\itemize{
\item \code{logical}, e.g. TRUE or FALSE
\item \code{function}. Needs as input an argument \code{row} which accepts a single row \code{tibble} and as output TRUE/FALSE.
}}

\item{canDeleteRow}{can be either of the following:
\itemize{
\item \code{logical}, e.g. TRUE or FALSE
\item \code{function}. Needs as input an argument \code{row} which accepts a single row \code{tibble} and as output TRUE/FALSE.
}}

\item{canCloneRow}{can be either of the following:
\itemize{
\item \code{logical}, e.g. TRUE or FALSE
\item \code{function}. Needs as input an argument \code{row} which accepts a single row \code{tibble} and as output TRUE/FALSE.
}}

\item{statusCol}{\code{character(1)} name of column with general status (e.g. modified or not).
if \code{NULL}, the data is interpreted as 'unmodified'.}
}
\value{
\code{character(1)} HTML
}
\description{
Create buttons to modify the row.
}
\details{
buttons used per row in the app.
}
