% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{createButtonsHTML}
\alias{createButtonsHTML}
\title{Helper function to write HTML}
\usage{
createButtonsHTML(suffix = "\%1$s", ns = "\%2$s")
}
\arguments{
\item{suffix}{\code{character(1)} sprintf placeholer for suffix}

\item{ns}{\code{character(1)} sprintf placeholder for ns}
}
\value{
\code{character(1)} HTML to be filled in with \code{sprintf}
}
\description{
Helper function to write HTML
}
\details{
generate HTML as character once and reuse.
Since buttons have to be generated a lot, this otherwhise slows down the app.
}
