% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDT.R
\name{createEditButtonHTML}
\alias{createEditButtonHTML}
\title{Generate HTML for an in-row edit button}
\usage{
createEditButtonHTML(ns, suffix, disabled = FALSE)
}
\arguments{
\item{ns}{\code{character(1)} namespace}

\item{suffix}{\code{character(1)} id of the row}

\item{disabled}{\code{logical(1)} wether or not the button has to be disabled}
}
\value{
\code{character(1)} HTML
}
\description{
Generate HTML for an in-row edit button
}
