% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoApp.R
\name{demoServer_DB}
\alias{demoServer_DB}
\title{Server of the DB demo app}
\usage{
demoServer_DB(id, conn)
}
\arguments{
\item{id}{\code{character(1)}}

\item{conn}{database connection object as given by \code{\link[DBI]{dbConnect}}.}
}
\value{
NULL, just executes the module server.
}
\description{
Server of the DB demo app
}
\author{
Jasper Schelfhout
}
