% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoApp.R
\name{demoUI_mtcars}
\alias{demoUI_mtcars}
\title{UI of the demo mtcars app}
\usage{
demoUI_mtcars(id)
}
\arguments{
\item{id}{\code{character(1)}}
}
\value{
HTML
}
\description{
UI of the demo mtcars app
}
\author{
Jasper Schelfhout
}
