% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devApp.R
\name{devServer}
\alias{devServer}
\title{Server of the development app}
\usage{
devServer(id, conn)
}
\arguments{
\item{id}{\code{character(1)}}

\item{conn}{database connection object as given by \code{\link[DBI]{dbConnect}}.}
}
\value{
NULL, just executes the module server.
}
\description{
Server of the development app
}
\author{
Jasper Schelfhout
}
