% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreignTbl.R
\name{fillDeductedColumns}
\alias{fillDeductedColumns}
\title{Fill data columns based on foreignTbls}
\usage{
fillDeductedColumns(tbl, foreignTbls)
}
\arguments{
\item{tbl}{\code{tbl}}

\item{foreignTbls}{list of foreign tbls as created by \code{\link{foreignTbl}}}
}
\value{
tbl
}
\description{
Fill data columns based on foreignTbls
}
\details{
When a combination of columns is not found in the foreignTbl,
fill the deductedColumns with NA.

on foreignTbls suggesting conflicting data,
an arbitrary choice is made. It is best to afterwards check with
checkForeignTbls to see if a valid result is obtained.
}
\author{
Jasper Schelfhout
}
