% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fixInteger64}
\alias{fixInteger64}
\title{Replace instances of integer64 with actual NA values instead of weird default 9218868437227407266}
\usage{
fixInteger64(x)
}
\arguments{
\item{x}{\code{data.frame}}
}
\value{
x with \code{integer64} columns set to \code{bit64::as.integer64(NA)}
}
\description{
Replace instances of integer64 with actual NA values instead of weird default 9218868437227407266
}
\details{
\href{https://github.com/Rdatatable/data.table/issues/4561}{github issue}
}
\author{
Jasper Schelfhout
}
