% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_dbi.R
\name{get_db_table_name}
\alias{get_db_table_name}
\title{Get name of the tbl in the database}
\usage{
get_db_table_name(x)
}
\arguments{
\item{x}{\code{tbl_dbi}}
}
\value{
SQL, the table name as used in the database
}
\description{
Get name of the tbl in the database
}
