% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInput.R
\name{inputServer.default}
\alias{inputServer.default}
\title{An input server for a data.frame}
\usage{
\method{inputServer}{default}(id, data)
}
\arguments{
\item{id}{character module id}

\item{data}{single row data.frame}
}
\value{
reactive modified version of data
}
\description{
An input server for a data.frame
}
\details{
Reads all inputs ids that are identical to column names of the data
and updates the data.
}
\author{
Jasper Schelfhout
}
