% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInput.R
\name{inputServer}
\alias{inputServer}
\title{An input server for a data.frame}
\usage{
inputServer(id, data, ...)
}
\arguments{
\item{id}{character module id}

\item{data}{single row data.frame}

\item{...}{further arguments for methods}
}
\value{
modified version of data
}
\description{
An input server for a data.frame
}
\author{
Jasper Schelfhout
}
