% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyInput.R
\name{selectInputDT_UI}
\alias{selectInputDT_UI}
\title{Create a DT select input}
\usage{
selectInputDT_UI(id)
}
\arguments{
\item{inputId}{}

\item{label}{}

\item{data}{}
}
\value{
HTML
}
\description{
Create a DT select input
}
\author{
Jasper Schelfhout
}
