% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectInputDT.R
\name{selectInputDT_UI}
\alias{selectInputDT_UI}
\title{Create a DT select input}
\usage{
selectInputDT_UI(id)
}
\arguments{
\item{id}{character}
}
\description{
Create a DT select input
}
\author{
Jasper Schelfhout
}
